feat(conhost): windowed console
This adds a third mode between the 'full' UI-focus console and minicon,
mainly to be used when dragging the console outside of the game window
or to have a full console always at hand even with UI focus not held.

Potential future improvements could include the option to hide the full
console portion of F8 (keeping the menu bar), or hiding the input field
if the console would be 'unclickable' (in the main game viewport and not
owning console focus).

(this was requested as a 'silly idea')