fix(ros): support third-party auth tokens
Rockstar started enforcing 'reauthentication' for third-party auth (such
as Steam or Epic) which broke RedM for users on those stores, but also
would break FiveM in that it would no longer auto-signin nor would the
license token be based on `steam:`.

Instead, we now defer to SCUI when we receive the '4000.210' error code
from ROS, which now also supports 'automatic' flows including the 'use
this account' prompt. On success, we then save the TpaToken and we use
it for any future requests.