fix(input/five): sniper zoom bug for scroll wheel
Since the mouse scroll delta stuff, sniper zoom with a mouse wheel would
either be zoomed all the way in, or all the way out, as apparently this
'other' value derived from dZ is a count of scroll intervals.

Technically this should be accumulated from partial scrolls still but to
fix the regression we just make it divide by 120 as integer again.

Smooth scrolling from the delta axis has still been confirmed to work.