fix(level-loader/five): no more city chunk if replacing map on b2699+
Since b2699, some parts of the base map would load even if replacing the
level meta with a custom one. This is because patchday27 has a changeset
("CCS_PATCHDAY27_NG_STREAMING_MAP") which updates a few map chunks such
as dt1, but which lacks a <associatedMap>MO_JIM_L11</associatedMap> tag.

This was reported on https://forum.cfx.re/t/5162859, but likely also in
some other places prior given it's a fairly old bug.