feat(onesyn/events): add ragdoll request event
Since some cheats have now an option to ragdoll any player server owners should be able to cancel that event.

My knowledge is limited, and I don't know how to figure out the event data. I put a placeholder so that someone with that knowledge should be able to replace that.