tweak(input/five): release cursor capture for the game's cursor
Custom bits of UI (NUI focus, as well as the F8 console/dear ImGui) do
this already, which makes the game's mouse cursor (such as in the pause
menu) behave differently.

This commit solves this by hooking up the QueryMayLockCursor event to a
flag specifying if the cursor is visible.