tweak(five): "add 2 additional crash fixes" for b3095
One of the functions patched in a0f35b50fd6ecb10f052341374fbe47eb9401764
changed quite significantly in b3095 - including being moved to a
different part of the audio update sequence.

This commit moves that specific audio cloth patch to its own unit and
updates it for b3095. Current theory behind the crash is a side-effect
of a vehicle fragment without a chassis_dummy bone.

Fixes #2448