build(server/five): fix linux server build by removing winapi dependency
Guard part of code that is in code/client/shared but not supposed to be used
on the server with #ifndef IS_FXSERVER. This is a hacky solution but
thats what we do in XBRInit.cpp as well. So it seems to be the standard
right now.

The long term solition would be to stop building code/client/shared on server
and move all dependencies from there (if there are any) to some shared place
between client and server. Maybe code/shared