tweak(game/five): remove dlc lists and use enforce game build instead
Stop using dlc block/allow lists that explicitly set list of dlcs to
load. Instead, when server specifies sv_enforceGameBuild - do not change
the actual build, but load appropriate set of dlcs.

This change should be transparent for users who never set dlc lists.

After this change IsGameBuildOrGreater type of checks are not needed
anymore. However this commit does not clean up those checks. This commit
is intended to go into a feature branch first. So by keeping those
checks we simplify this iteration and avoid painful merge conflicts in
the future. Since work on title update is being done in parralel. The
cleanup will be done after testing is complete.

Note: I am not sure when do we want to show the actual build and when
the one that we are pretending to be. Currently in all visible for users
places it is set to be the pretended / requested by server build.