Update LoadResourceFile.md
Modified the parameters from fileName to directroy because the parameter is the path and not the name, also changed the descritpton from "The name of the file" to "The path of the file in the resource". The example has been modified too, now the example shows that the "data.json" file is in a folder called client.