fix(game/five): load correct files from update.rpf
After switching to blocking DLCs instead of loading old build completely
we have noticed that in some areas NPCs are placed in the air. This is
because their configuration is taken from
update.rpf/x64/levels/gta5/scenario folder. And the update.rpf is loaded
for the actual build - not the one we are pretending to be.

To load correct versions of these and some other files we mount an
additional .rpf archive that overrides those files.

Also we now don't need the original DLC blocklisting logic as we can
just redirect the game to read the DLC list from update.rpf that
corresponds to the older game build.