fix(game/five): set correct server side game build on new build system
This is a followup fix for:
https://github.com/citizenfx/fivem/commit/df68640468051893a722a96aff90cb35412c186d

In ServerGameState we were relying on GTA_FIVE and IS_RDR3 macros which
are not defined on the server side. This commit fixes that depencency.

I have not encountered any issues caused by this myself, but it could
potentially break things in some unexpected places.