fix(github/ci): restrict permissions and fix PR checkout issue
The previous checkout behavior incorrectly fetched the upstream master branch instead of the PR's repository and branch, resulting in untested changes from PRs. This update corrects that issue, ensuring PR changes are properly tested. Additionally, we have restricted the Action Run's permissions to reduce the risk of unauthorized requests in the future.