fix(streaming/five): improve archetype handling
Although the logic in 0e05dbbc0fe962b45b1748911988d8d27b8735c6 was likely correct, it didn't address a separate issue, which is that LookupModelId retrieves an archetype's model id though the ArchetypeMap.

When there are duplicate archetypes, this breaks stuff, because that model id doesn't necessarily correspond to the provided archetype.
Instead, just cache the model id inside of the archetype and return that.

Also, protect archetypes used by scripted objects, to ensure they are not prematurely unloaded.