feat(game/debug): skip pool size increase validation id dev mode
If moo 31337 is set - skip pool size validation. It should be set
independently on server and client side to work. If only set on the
server side - the client will not be able to connect if the pool size
increase is invalid.

This change is done because multiple server owners asked for a way to
debug their servers with different experimantal pool sizes without having
to wait for CFX team to add them to allow list.