fix(game/cli): correctly parse cli arguments on game startup
Previously we searched for a substring in the whole CLI string, which
included path to the executable. This could cause issues if the
executable path contains substring that can be mistaken for the pure
level or requested build flag.

We did not get reports about this issue earlier because substrings like
b1604 are probably not so common in file pathes. But after we added
build one (b1) in
https://github.com/citizenfx/fivem/commit/df68640468051893a722a96aff90cb35412c186d -
some people reported that they are stuck in infinite reload loop.

This commit makes it so we parse CLI arguments first and then search for
flags among the parsed arguments.

There are more places besides pure level and build flags were we work with
cli incorrectly. But fixing them all requires a large scale refactoring.
We may do it in a followup commits.