fix(editor/five): revert streaming cleanup on clip reload
Revert
https://github.com/citizenfx/fivem/commit/d18f651833730b2ce0ebc2086282b61395c1f88e
. The original fix was supposed to fix a crash when editing clips
recorded on servers with different set of resources. In order to do so
we initiated a streaming cleanup.

It did fix the issues during local testing. But seems like it broke
multuple users as well. Context:
https://github.com/citizenfx/fivem/issues/2846#issuecomment-2489199697

It is not clear why the fix causes the issues it causes. So we should
revert it and and do the investigation afterwards.