tweak(mumble): don't recreate the UDP handle whenever the client stops receiving UDP packets
This looks to have been needed for an old issue where FiveM's mumble implementation didn't properly handle resetting the sockets on stream close, which is handled properly now.

Mumble hits a "fast path" for reconnecting a client if it reuses the same socket, ad it can just find the client off of the socket and reset the clients crypt state

Whenever we recreate the socket we instead hit the "slow path" where the server has to iterate over *every* client connected to mumble and try to see if any of them can decrypt the UDP packet that was sent, which is not optimal.