fix(game/five): fix weapon spin up sync if onesync is turned on
Before when spinning up a weapon - the spin up would happen much slower
on remote client. This issue only appeared when onesync is turned on.
The problem was more prominent with some weapons like heavy machine gun.

The reason for the issue is that when
CTaskMotionAiming::ComputePitchSignal is called on remote client, the
weapon task target is not yet updated. So the first ComputePitchSignal
call ends up setting current pitch to either nearly 0 (weapon facing
down) or 0.5 (weaponf facing forward). And the consequent
ComputePitchSignal calls would not apply the pitch change directly but
rather smooth out the transition. This causes the synchronisation issue.

When onesync is turned off the issue is not present because we end up
calling aiTaskTree::UpdateTasks more often and are just in time to
update the weapon task target before ComputePitchSignal is called for
the first time.

Thi change removes pitch smoothing out for remote client to make sure
that the pitch always equals the one on the source client.

Alternative fix would be to try running UpdateTasks more often, but it
is harder to do and could cause more unexpected side effects.