fix(core/rdr3): AMV hardcoded limit causing blinking
There is a global bitset used to store the enabled/disabled state of
every AMV. This bitset has a hardcoded size of 6016 entries, once there
are more AMVs than the bitset size, it starts reading/writing
out-of-bounds. This causes the AMVs to start blinking and can
potentially crash.

This patch re-allocates that bitset with enough space to fit the
capacity of all the AMV pools. The functions that access it are patched
to point to the new buffer and use the new size constants.