fix(five/sync): make sure all clients have the same game build
Follow up fix for
https://github.com/citizenfx/fivem/commit/3850efdceb72f1d93c076567ec5f04b4b5edf81b

That change was made under the assumption that
sv_replaceExeToSwitchBuilds is always sent by the server. However the
older server builds don't do that. So we revert the part of commit which
changed default sv_replaceExeToSwitchBuilds on client side.

After this fix sv_replaceExeToSwitchBuilds will still be set to false by
default on the latest server builds.

If client side default value will be changed in the future - we should
push it to all channels at the same time.