feat(gta-streaming-five): patch sky portal flags
As Iridium said the at `83 B9 ? ? ? ? ? 4C 8B CA 4C 8B C1` function needed to be patched because the size of the skyPortalFlags was the size of the FW_PORTAL_SCENE_GRAPH_NODE pool. Since we doubled it(800) that size was not updated.
I created a function that returns the address of the new skyPortalFlag with its correct size based on the address of the class self, so if there is another instance of the same class it will have a different skyPortalFlags.
I patched the AddSkyPortal function to use the new SkyPortalFlag array and the same with the SetIsSkyPortalVisInterior.
Fixed the offset of the pool size constructor of FW_FIXED_ENTITY_CONTAINER.