fix(gta-streaming-five): SetIsSkyPortalVisInterior with wrong registers values
Previously I was storing the address of `skyPortalFlagArray` in rcx, which was not correct because later in `mov(dword_ptr[rcx+0x378], r9d);` rcx was used to set `m_skyPortalCount` to 0, since having the address of skyPortalFlagArray as rcx this was not working correctly. Now we are storing the value of the original code in rcx.
I have removed `add(rcx, 0x37C);` because it is the offset to access the `skyPortalFlagArray` of the class but since we are allocating it in the c++ code, the only thing that offset was doing was jumping over a memory hole that was not used.
Finally after setting `m_skyPortalCount` to 0 the value of skyPortalFlagArray is stored in rcx to later call `memset`.