fix(net/sbac): don't duplicate the registration of the state bag component
This gets called further down in `rage::netObjectMgr::RegisterNetworkObject(obj)`
which is hooked to call `CloneManagerLocal::RegisterNetworkObject` which
will *then* register the state bag in a queued mode where it will
wait for the client to ack a packet from the server for the object
before it would apply it.

Calling it here could lead to the client still getting state bag changes
for this even if the entity didn't *actually* exist for the client.