fix: use actual gta exe build in fivem process name
Before we used the requested build numer: `GetRequestedGameBuild`
(i.e. one that is being mimicked with dlc set). It was done
intentionally a while ago to make the fact that we are running different
game build as unnoticeable as possible. But the implications of this on
symbol server were not taken into account. As a result sometimes we
wouldn't be able to load symbols due to name mismatch.