tweak(net): generate truly random "GUID"'s
This is old code that only gets used on initial client connection in cases
where the peer doesn't match the same TCP peer (for situations like Cloudflare Warp or other VPN's)

This ID previously allowed for someone to get disconnected if they had the
unfortunate coincidence of having the same tick time as another player on
initial join, and while this happened *rarely* it still happened.

This makes the GUID a "more random" number, and in situations where the GUID
will collide it will drop the connecting client instead of the existing client.