tweak(core/five): make it optional to disable component validation checks
Setting null drawables/invalid variations is sometimes used in scripts that make screenshots of clothing pieces, for example. With the validation in place, the head(or other components) cannot be invisible anymore, which breaks these screenshots, so making this validation deactivatable is crucial for these scripts to work. It also never crashed the game for me before the validation.