fix(streaming/rdr3): don't attempt to reload collisions
In RDR3 g_lastStreamingName was set to the "extension" from GetFileTypeIdFromExtension. For most custom assets this caused g_lastStreamingName to contain garbage data. Leading to functions like GetStreamingIndexForName or GetStreamingNameForIndex to not work as intended.

In #3543 this behaviour was corrected in order to fix GetStreamingIndexForName for custom fonts, however this unintentionally caused ReloadMapStore to not fail at GetStreamingIndexForName revealing an issue with reloading custom collision files in the RDR3 streaming implementation. Leading to them failing in game and cause crashes. This commit disables this behaviour again while fixing GetStreamingIndexForName and related functions to ensure compatability with #3543.