feat(net/rdr3): automatic stack resizing
Instead of manually patching a function stacks and its stack usage within the function. It's been replaced with an efficient helper function using udis86 to find all stack usage and replace it with the new stack sizes. This makes stack resizing quicker and should make it easier to update to newer executable versions.