tweak(net/rdr3): disable voice chat
This is a temporary and mostly experimental changes, the built in voice chat makes use extensive use of <32 sized arrays that may be causing issues at >33 players. The previous attempt to disable it with mumble voip support for RedM doesn't seem to account for everything, so for now we no-op the init and update functions.