feat(scripting/five): Prevent the pause menu from creating any script threads.
In Singleplayer and GTAO, certain pause menu tabs may stream in and start additional scripts/threads to fill in dynamic content. For example to handle clicking on race/deathmatch join blips, or fill in the mission objective in the GAME tab.

However bad script containing calls such as `TerminateThread(0)` as described in #3079, often end up killing these scripts, causing the game to crash the next time it's paused. Since this functionality is completely unused in FiveM, might as well patch it out.