tweak(native/codegen): only check that we're not server before filtering
Currently if we set server natives to be for specific games (like `gta5` or `rdr3`)
it will be filtered out of server codegen because the default game value is 'gta5'
when doing server codegen.

On the server all server natives exist at all times, so we just adjust the native filtering
to respect this.