tweak(mumble): reduce locking around GetTalkers
- `_isAnyoneTalking` gets called a decent bit in game code, and for some
reason the code here iterates *the entire talker list* again, and this can
presummably happen multiple times a frame and cause lock contention. Most
of the work here is already handled by another call that happens per frame
so we juse reuse `g_talkers` and `o_talkers`.

- `GetTalkers` also did a bunch of un-needed locking on every `GetUser`
call when it was trying to get users from their session id's, so we change
how we expect data to be handed back down and add a new function to handle
getting all of the user names and reduce the amount of locking/unlocking
we do.