feat(script-runtimes/v8): add memory info per context
This uses the experimental `MeasureMemory` api to measure all of the context's
used memory and stores the amount of used memory into the global `g_memoryInfo`.

This should add a decent bit of debugability to existing scripts that use V8, and hopefully
help script devs figure out if they're using a lot of memory.

In the future it might make sense to have a "common" v8 class that we can use to
share structure between the new Node ScRT and the v8 ScRT to reduce the duplication
here.

This also cleans up some existing code by using the `CfxV8EmbeddedData` enum to better
describe what certain embedder data does.