feat(scrt/v8): add support for source maps
This uses a non-minified version of https://github.com/7rulnik/source-map-js
to handle loading source maps.

This also changes behavior for the Node ScRT to go back to using the shared
v8 unhandled promise rejection so we get proper source maps for these calls.

This adjust the behavior of FORMAT_STACK_TRACE so we don't eat errors if they
happen on the same tick for JS/C# at the cost that we get less information
than we normaly would, but this should only really happen with export calls,
in that case we only lose information from Lua ScRT's since JS/C# both will
supply their current stack trace as arguments.